test: split test_particle_collection io and data
